把提示工具换成了clangd，完成了几个TODO

+++ include/sys/types.h
@@ -22,7 +22,7 @@ typedef int mode_t;
 typedef unsigned long long off_t;
 typedef int uid_t;
 typedef int nlink_t;
-typedef unsigned long size_t,useconds_t;
+typedef unsigned long long size_t,useconds_t;
 typedef long ssize_t;
